refactor(backend): extract RoutineScheduler from AgentService
The routine domain — the rows, their runs, the six `openbot` routine tools
and the single timer that fires the next due one — moves to
`src/backend/agent/routine-scheduler.ts`. The six public methods stay on the
facade as delegations, so the surface is unchanged.

`AgentRoutineStore` is now constructed by the scheduler rather than by the
service, which is what makes the boundary real: the run status transitions,
the conversation events they append, and the timer that produces them can no
longer be driven from anywhere else.

Two things the scheduler now owns outright, that were previously spread:

- **The `RoutineAttention` port PR 4 defined.** `AttentionRegistry` receives
  `this.#routines` directly instead of an inline adapter — a routine run
  blocked on a question is `needs-attention`, and answering returns it to
  `running`, which is how a user tells a stalled routine from a working one.
- **Its clause in the drain mute registry.** `#routineDeletionBots` becomes
  the scheduler's own `#deletionBots` and both `#scheduleDrain` and
  `#drainBot` now ask `this.#routines.mayDrain(botId)`. This is the
  incremental half of the plan's most valuable structural change — each PR
  converts its own guard clause — not scope creep. Only PR 6's duplication
  clause is left inline.

`RoutineHooks` is deliberately all writes: the reads go through `store`,
`mailbox` and `conversation` directly, so the hook list is exactly the set of
domains a routine mutates and nothing more. `awaitDrain` and `interrupt` are
there because a deletion must stop live runs before removing their routine;
`pendingDuplicateBots` because a half-copied agent must not fire.

The leak the plan predicted stays open and is not papered over:
`#emitQueue` is queue-engine code and remains in the service, so it keeps a
`this.#routines.reconcileDelivery(delivery): boolean` call and raises one
`routines-changed` per bot rather than per delivery. Inverting that with an
event bus is ruled out — `team-api-server.ts` rebroadcasts every event
unfiltered onto the released Team API wire protocol.

Verification, test files untouched: all five agent-service shards and the
five `src/backend/agent/*.test.ts` files pass (93 tests),
`bun run typecheck:node` is clean, the private-member delta is exactly the
fifteen removed names with none added, and `BEGIN IMMEDIATE` is still one
occurrence under `src/backend/agent*`.

Delta: −465 in `agent-service.ts` (3140 → 2675), +629 in the new file,
against the plan's −405 / +470.
